requirements.txt
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0

# MongoDB
pymongo==4.6.0
motor==3.3.2
dnspython==2.4.2

# Environment and CORS
python-dotenv==1.0.0

# ML Libraries (must match ml_models requirements)
darts[all]==0.27.0
lightgbm==4.0.0
pandas==2.0.0
numpy==1.24.0
scikit-learn==1.3.0
joblib==1.3.0
python-dateutil==2.8.2

# Additional dependencies from ml_models
yfinance>=0.2.28
ta>=0.11.0
matplotlib>=3.7.0
plotly>=5.17.0

# HTTP client (optional, for testing)
httpx==0.25.2
